move parse support to another endpoint file simplify calls and loading of records
